simplify implementation and code reuse docs
